refactor: move valueHint translation to the renderers
