+++ packages/enmeshed_types/pubspec.yaml
@@ -10,7 +10,6 @@ environment:
   sdk: ^3.2.0
 
 dependencies:
-  collection: ^1.18.0
   equatable: ^2.0.5
   json_annotation: ^4.8.1
   meta: any
