+++ packages/renderers/lib/src/attribute/complex_attribute_list_tile.dart
@@ -2,6 +2,8 @@ import 'package:enmeshed_types/enmeshed_types.dart';
 import 'package:flutter/material.dart';
 import 'package:translated_text/translated_text.dart';
 
+import 'value_hint_translation.dart';
+
 class ComplexAttributeListTile extends StatelessWidget {
   final String title;
   final List<({String label, String value})> fields;
