+++ packages/renderers/lib/src/attribute/widgets/street_address_attribute_renderer.dart
@@ -2,6 +2,8 @@ import 'package:enmeshed_types/enmeshed_types.dart';
 import 'package:flutter/material.dart';
 import 'package:translated_text/translated_text.dart';
 
+import '../value_hint_translation.dart';
+
 class StreetAddressAttributeRenderer extends StatelessWidget {
   final StreetAddressAttributeValue value;
   final ValueHints valueHints;
